// projects/orunsurvey/src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      survey_questions: {
        Row: {
          created_at: string
          id: string
          is_required: boolean
          media_url: string | null
          options: Json | null
          question_order: number
          question_text: string
          question_type: string
          survey_id: string
          time_settings: Json | null
        }
        Insert: {
          created_at?: string
          id?: string
          is_required?: boolean
          media_url?: string | null
          options?: Json | null
          question_order: number
          question_text: string
          question_type: string
          survey_id: string
          time_settings?: Json | null
        }
        Update: {
          created_at?: string
          id?: string
          is_required?: boolean
          media_url?: string | null
          options?: Json | null
          question_order?: number
          question_text?: string
          question_type?: string
          survey_id?: string
          time_settings?: Json | null
        }
        Relationships: [
          {
            foreignKeyName: "survey_questions_survey_id_fkey"
            columns: ["survey_id"]
            isOneToOne: false
            referencedRelation: "surveys"
            referencedColumns: ["id"]
          },
        ]
      }
      survey_responses: {
        Row: {
          "7c49fdca-ceb9-4edf-abc3-a43f7d8189a4": Json | null
          "88aec870-d8e9-4a1c-a56b-2891ee23c4c9": string | null
          "8ded2532-4d0e-4afe-8ad8-bfbb1ec6542c": string | null
          additional_comments: string | null
          created_at: string
          "e89e505d-027e-4a75-8240-d1ef53ef849a": string | null
          exam_type: string
          id: string
          join_class: string
          join_class_other: string | null
          name: string
          school: string
          survey_id: string | null
          time_slots: Json
        }
        Insert: {
          "7c49fdca-ceb9-4edf-abc3-a43f7d8189a4"?: Json | null
          "88aec870-d8e9-4a1c-a56b-2891ee23c4c9"?: string | null
          "8ded2532-4d0e-4afe-8ad8-bfbb1ec6542c"?: string | null
          additional_comments?: string | null
          created_at?: string
          "e89e505d-027e-4a75-8240-d1ef53ef849a"?: string | null
          exam_type: string
          id?: string
          join_class: string
          join_class_other?: string | null
          name: string
          school: string
          survey_id?: string | null
          time_slots?: Json
        }
        Update: {
          "7c49fdca-ceb9-4edf-abc3-a43f7d8189a4"?: Json | null
          "88aec870-d8e9-4a1c-a56b-2891ee23c4c9"?: string | null
          "8ded2532-4d0e-4afe-8ad8-bfbb1ec6542c"?: string | null
          additional_comments?: string | null
          created_at?: string
          "e89e505d-027e-4a75-8240-d1ef53ef849a"?: string | null
          exam_type?: string
          id?: string
          join_class?: string
          join_class_other?: string | null
          name?: string
          school?: string
          survey_id?: string | null
          time_slots?: Json
        }
        Relationships: [
          {
            foreignKeyName: "survey_responses_survey_id_fkey"
            columns: ["survey_id"]
            isOneToOne: false
            referencedRelation: "surveys"
            referencedColumns: ["id"]
          },
        ]
      }
      surveys: {
        Row: {
          basic_fields: Json | null
          created_at: string
          created_by: string | null
          description: string | null
          id: string
          intro_text: string | null
          is_active: boolean
          slug: string
          title: string
          updated_at: string
        }
        Insert: {
          basic_fields?: Json | null
          created_at?: string
          created_by?: string | null
          description?: string | null
          id?: string
          intro_text?: string | null
          is_active?: boolean
          slug: string
          title: string
          updated_at?: string
        }
        Update: {
          basic_fields?: Json | null
          created_at?: string
          created_by?: string | null
          description?: string | null
          id?: string
          intro_text?: string | null
          is_active?: boolean
          slug?: string
          title?: string
          updated_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never